chore(http): delete deprecated mcp/ and agent/http/ modules (Phase 6)
- Delete apps/server/src/mcp/server.ts and index.ts (replaced by http/routes/mcp.ts)
- Delete apps/server/src/agent/http/HttpServer.ts, types.ts, index.ts (replaced by http/)
- Move ChatRequestSchema and related types to http/types.ts
- Update imports in GeminiAgent.ts, agent/types.ts, agent/index.ts
- Remove deprecated exports from agent/index.ts
- Remove commented out startMcpServer and startAgentServer functions from main.ts
