--- apps/server/src/agent/http/index.ts
@@ -1,12 +1,0 @@
-/**
- * @license
- * Copyright 2025 BrowserOS
- * SPDX-License-Identifier: AGPL-3.0-or-later
- */
-export { createHttpServer } from './HttpServer.js'
-export type {
-  ChatRequest,
-  HttpServerConfig,
-  ValidatedHttpServerConfig,
-} from './types.js'
-export { ChatRequestSchema, HttpServerConfigSchema } from './types.js'
