+++ apps/server/src/http/types.ts
@@ -4,10 +4,48 @@
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
+import { z } from 'zod'
+import { VercelAIConfigSchema } from '../agent/agent/gemini-vercel-sdk-adapter/types.js'
 import type { RateLimiter } from '../agent/rate-limiter/index.js'
 import type { Logger, McpContext, Mutex } from '../common/index.js'
 import type { ControllerContext } from '../controller-server/index.js'
 import type { ToolDefinition } from '../tools/index.js'
+
+// Chat request schemas (moved from agent/http/types.ts)
+export const TabSchema = z.object({
+  id: z.number(),
+  url: z.string().optional(),
+  title: z.string().optional(),
+})
+
+export type Tab = z.infer<typeof TabSchema>
+
+export const CustomMcpServerSchema = z.object({
+  name: z.string(),
+  url: z.string().url(),
+})
+
+export type CustomMcpServer = z.infer<typeof CustomMcpServerSchema>
+
+export const BrowserContextSchema = z.object({
+  windowId: z.number().optional(),
+  activeTab: TabSchema.optional(),
+  selectedTabs: z.array(TabSchema).optional(),
+  tabs: z.array(TabSchema).optional(),
+  enabledMcpServers: z.array(z.string()).optional(),
+  customMcpServers: z.array(CustomMcpServerSchema).optional(),
+})
+
+export type BrowserContext = z.infer<typeof BrowserContextSchema>
+
+export const ChatRequestSchema = VercelAIConfigSchema.extend({
+  conversationId: z.string().uuid(),
+  message: z.string().min(1, 'Message cannot be empty'),
+  contextWindowSize: z.number().optional(),
+  browserContext: BrowserContextSchema.optional(),
+})
+
+export type ChatRequest = z.infer<typeof ChatRequestSchema>
 
 /**
  * Hono environment bindings for Bun.serve integration.
