--- apps/server/src/mcp/index.ts
@@ -1,12 +1,0 @@
-/**
- * @license
- * Copyright 2025 BrowserOS
- *
- * MCP Server Package - Provides Model Context Protocol server implementation
- */
-
-export {
-  createHttpMcpServer,
-  type McpServerConfig,
-  shutdownMcpServer,
-} from './server.js'
